+++ src/main/java/com/sparta/server/threeserving/order_management/dto/response/DashboardTrendResponse.java
@@ -25,8 +25,8 @@ public class DashboardTrendResponse {
         private Long rejectedCount;
         private Long canceledCount;
         private Long totalSalesAmount;
-        private BigDecimal avgCookTime;
-        private BigDecimal acceptRate;
+        private Double avgCookTime;
+        private Double acceptRate;
 
     }
 }
